+++ CHANGELOG.md
@@ -1,14 +1,13 @@
 
 ## [Unreleased]
 
-- [ ] adition of actual parsing errors
-- [ ] removal of blatantly duplication
 - [x] move parser entities to their own files
-- [ ] simulator
+- [ ] adition of actual parsing errors
 - [ ] human friendly (even more) error messages
     - [ ] introduciton of a error object within the execeptions
           with the line and column information, but also with
           suggestions of how to correct the problem when possible.
+- [ ] simulator
 
 ## [0.1.0] - 2023-09-01
